+++ peak-blog/src/components/blog/post/post-preview.tsx
@@ -5,8 +5,7 @@ import {Node} from "slate";
 import Link from "next/link";
 import moment from "moment";
 import {EstimateReadTime} from "./read-time/EstimatedReadTime";
-import Image from "next/image";
-import {ImageProps} from "next/dist/client/image";
+import {ConditionalImageLoader} from "../../primitives/image/ConditionalImageLoader";
 
 export const BlogPostPreview = (props: { post: PeakPost }) => {
     const { post } = props
@@ -18,7 +17,7 @@ export const BlogPostPreview = (props: { post: PeakPost }) => {
 
     return (
         <div className={"mb-20 flex min-h-150"}>
-            <ImagePreview post={post}/>
+            <ConditionalImageLoader src={post.cover_image} width={"200px"} height={"150px"} layout={"intrinsic"} className={"mr-5"}/>
             <div className="flex flex-col flex-grow">
                 <Link href={`post/${post.id}`}>
                     <h2 className={"mb-2 cursor-pointer hover:text-blue-500 font-semibold text-3xl"}>{title}</h2>
@@ -32,25 +31,4 @@ export const BlogPostPreview = (props: { post: PeakPost }) => {
             </div>
         </div>
     )
-}
-
-const ImagePreview = (props: { post: PeakPost }) => {
-    const cover_image = props.post.cover_image
-    return (!cover_image) ? null : <div className={"mr-5"}>
-        <NextImageWrapper src={cover_image} width={"200px"} height={"125px"} layout={"intrinsic"} alt={"Image Preview"}/>
-    </div>
-}
-
-const NextImageWrapper = (props: ImageProps) => {
-    // @ts-ignore
-    const { src, className, alt, width, layout, height } = props
-    return (src.startsWith("https://storage.googleapis.com")) ?
-        <Image
-            src={src}
-            alt={alt}
-            width={width}
-            layout={layout}
-            height={height}
-        /> : <img src={src} className={props.className} alt={alt} style={{width: width, height: height}}/>
-
 }
