+++ peak-blog/src/components/blog/post/post.tsx
@@ -3,10 +3,10 @@ import {PeakPost} from "component-library";
 import {DisplayEditor} from "../../rich-text-editor/DisplayEditor";
 import { Node } from "slate";
 import moment from "moment";
-import Link from "next/link";
 import {useAppContext} from "../../../data/context";
 import styles from "../../../../styles/Home.module.css";
 import {EstimateReadTime} from "./read-time/EstimatedReadTime";
+import {ConditionalImageLoader} from "../../primitives/image/ConditionalImageLoader";
 
 export const BlogPost = (props: { post: PeakPost }) => {
     const { post } = props
@@ -18,12 +18,12 @@ export const BlogPost = (props: { post: PeakPost }) => {
 
     return (
         <div className={styles.postContainer}>
-            <Link href={`post/${post.id}`}>
-                <h1 className={"mb-8 cursor-pointer"}>{title}</h1>
-            </Link>
-            <div className={"text-gray-500 font-medium text-sm mb-14"}>
+            <h1 className={"mb-2 text-4xl"}>{title}</h1>
+            <h2 className={"text-gray-500 text-xl font-light leading-normal mb-4"}>{post.subtitle}</h2>
+            <div className={"text-gray-500 text-sm mb-10"}>
                 <span className={"mr-1"}>{author.given_name} {author.family_name}</span> • <span className={"mx-1"}>{moment(post.created_at).format('LL') }</span> • <EstimateReadTime className={"mx-1"} body={post.body}/>
             </div>
+            <ConditionalImageLoader src={post.cover_image} width={"100%"} height={"auto"} layout={"responsive"}/>
             <DisplayEditor value={bodySanTitle} postId={post.id}/>
         </div>
     )
